fix: Delete item from list

+++ app/listas/info/include/dItem.php
@@ -10,12 +10,14 @@ $item = filter_input_array(INPUT_POST, [
     "list" => FILTER_SANITIZE_NUMBER_INT,
     "id" => FILTER_SANITIZE_NUMBER_INT
 ]);
-
 $id = $item['id'];
-$list = $item['list'];
-$deleteItem = deleteItem($id);
-if(!$deleteItem['success']){
-    $alert = "&alert=1";
+$idList = $item['list'];
+$list = getList(null, $idList)[0][5];
+if($list){
+    $deleteItem = deleteItem($id);
+    if(!$deleteItem['success']){
+        $alert = "&alert=1";
+    }
 }
 
-header("Location: ../?item=$list&msg=".$deleteItem['message']. $alert);
+header("Location: ../?list=$list&msg=".$deleteItem['message']. $alert);
